requirements.txt
tensorflow
numpy
matplotlib
pillow
opencv-python
streamlit